# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyTree Enviros API",
        "description": "API documentation for MyTree Enviros - Tree Sponsorship, Adoption & E-commerce Platform",
        "contact": {
            "name": "API Support",
            "email": "support@mytreeenviros.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://mytreeenviros.com/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.mytreeenviros.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/sliders": {
            "get": {
                "tags": [
                    "Sliders"
                ],
                "summary": "List sliders",
                "description": "Returns a list of sliders for the homepage. Active-only by default.",
                "operationId": "listSliders",
                "parameters": [
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Filter by active status (true/false). Defaults to true.",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit the number of returned sliders (1-100).",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Sort order by ID (asc or desc). Defaults to desc.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Slider"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/sliders/{id}": {
            "get": {
                "tags": [
                    "Sliders"
                ],
                "summary": "Get slider by ID",
                "description": "Show a single slider by ID.",
                "operationId": "getSlider",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Slider ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Slider"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Slider": {
                "title": "Slider",
                "description": "Homepage slider model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Plant More Trees",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Join our green mission and plant a tree today.",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "main_image_url": {
                        "description": "Signed URL to access the slider image (expires in 60 minutes)",
                        "type": "string",
                        "format": "uri",
                        "example": "https://api.example.com/media/12345?expires=1700000000&signature=xyz",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-02T12:34:56.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}